move/rename SolutionTrackers to new model_evaluation module
