+++ algorithms/indexing/indexer.py
@@ -1267,10 +1267,12 @@ class indexer_base(object):
 
     def choose_best_orientation_matrix(self, candidate_orientation_matrices):
 
+        from dials.algorithms.indexing import model_evaluation
+
         solution_scorer = self.params.basis_vector_combinations.solution_scorer
         if solution_scorer == "weighted":
             weighted_params = self.params.basis_vector_combinations.weighted
-            solutions = SolutionTrackerWeighted(
+            solutions = model_evaluation.ModelRankWeighted(
                 power=weighted_params.power,
                 volume_weight=weighted_params.volume_weight,
                 n_indexed_weight=weighted_params.n_indexed_weight,
@@ -1278,7 +1280,7 @@ class indexer_base(object):
             )
         else:
             filter_params = self.params.basis_vector_combinations.filter
-            solutions = SolutionTrackerFilter(
+            solutions = model_evaluation.ModelRankFilter(
                 check_doubled_cell=filter_params.check_doubled_cell,
                 likelihood_cutoff=filter_params.likelihood_cutoff,
                 volume_cutoff=filter_params.volume_cutoff,
@@ -1347,7 +1349,7 @@ class indexer_base(object):
                 rmsds = refiner.rmsds()
                 xy_rmsds = math.sqrt(rmsds[0] ** 2 + rmsds[1] ** 2)
                 model_likelihood = 1.0 - xy_rmsds
-                soln = Solution(
+                soln = model_evaluation.Result(
                     model_likelihood=model_likelihood,
                     crystal=experiments.crystals()[0],
                     rmsds=rmsds,
@@ -1489,11 +1491,11 @@ class indexer_base(object):
         if len(solutions):
             logger.info("Candidate solutions:")
             logger.info(str(solutions))
-            best_solution = solutions.best_solution()
-            logger.debug("best model_likelihood: %.2f" % best_solution.model_likelihood)
-            logger.debug("best n_indexed: %i" % best_solution.n_indexed)
-            self.hkl_offset = best_solution.hkl_offset
-            return best_solution.crystal, best_solution.n_indexed
+            best_model = solutions.best_model()
+            logger.debug("best model_likelihood: %.2f" % best_model.model_likelihood)
+            logger.debug("best n_indexed: %i" % best_model.n_indexed)
+            self.hkl_offset = best_model.hkl_offset
+            return best_model.crystal, best_model.n_indexed
         else:
             return None, None
 
@@ -1754,281 +1756,6 @@ class indexer_base(object):
 
     def find_lattices(self):
         raise NotImplementedError()
-
-
-from libtbx import group_args
-
-
-class Solution(group_args):
-    pass
-
-
-def filter_doubled_cell(solutions):
-    accepted_solutions = []
-    for i1, s1 in enumerate(solutions):
-        doubled_cell = False
-        for (m1, m2, m3) in (
-            (2, 1, 1),
-            (1, 2, 1),
-            (1, 1, 2),
-            (2, 2, 1),
-            (2, 1, 2),
-            (1, 2, 2),
-            (2, 2, 2),
-        ):
-            if doubled_cell:
-                break
-            a, b, c = s1.crystal.get_real_space_vectors()
-            new_cryst = Crystal(
-                real_space_a=1 / m1 * a,
-                real_space_b=1 / m2 * b,
-                real_space_c=1 / m3 * c,
-                space_group=s1.crystal.get_space_group(),
-            )
-            new_unit_cell = new_cryst.get_unit_cell()
-            for s2 in solutions:
-                if s2 is s1:
-                    continue
-                if new_unit_cell.is_similar_to(
-                    s2.crystal.get_unit_cell(), relative_length_tolerance=0.05
-                ):
-                    R, axis, angle, cb = difference_rotation_matrix_axis_angle(
-                        new_cryst, s2.crystal
-                    )
-                    if (angle < 1) and (s1.n_indexed < (1.1 * s2.n_indexed)):
-                        doubled_cell = True
-                        break
-
-        if not doubled_cell:
-            accepted_solutions.append(s1)
-
-    return accepted_solutions
-
-
-# Tracker for solutions based on code in rstbx/dps_core/basis_choice.py
-class SolutionTrackerFilter(object):
-    def __init__(
-        self,
-        check_doubled_cell=True,
-        likelihood_cutoff=0.8,
-        volume_cutoff=1.25,
-        n_indexed_cutoff=0.9,
-    ):
-        self.check_doubled_cell = check_doubled_cell
-        self.likelihood_cutoff = likelihood_cutoff
-        self.volume_cutoff = volume_cutoff
-        self.n_indexed_cutoff = n_indexed_cutoff
-        self.all_solutions = []
-        self.filtered_solutions = []
-
-    def append(self, item):
-        self.all_solutions.append(item)
-        self.update_analysis()
-
-    def __len__(self):
-        return len(self.filtered_solutions)
-
-    def filter_by_likelihood(self, solutions):
-        best_likelihood = max(s.model_likelihood for s in solutions)
-        offset = 0
-        while (best_likelihood + offset) <= 0:
-            offset += 1
-        return [
-            s
-            for s in solutions
-            if (s.model_likelihood + offset)
-            >= (self.likelihood_cutoff * (best_likelihood + offset))
-        ]
-
-    def filter_by_volume(self, solutions):
-        # filter by volume - prefer solutions with a smaller unit cell
-        min_volume = min(s.crystal.get_unit_cell().volume() for s in solutions)
-        return [
-            s
-            for s in solutions
-            if s.crystal.get_unit_cell().volume() < (self.volume_cutoff * min_volume)
-        ]
-
-    def filter_by_n_indexed(self, solutions, n_indexed_cutoff=None):
-        if n_indexed_cutoff is None:
-            n_indexed_cutoff = self.n_indexed_cutoff
-        # filter by number of indexed reflections - prefer solutions that
-        # account for more of the diffracted spots
-        max_n_indexed = max(s.n_indexed for s in solutions)
-        return [s for s in solutions if s.n_indexed >= n_indexed_cutoff * max_n_indexed]
-
-    def update_analysis(self):
-        # pre-filter out solutions that only account for a very small
-        # percentage of the indexed spots relative to the best one
-        self.filtered_solutions = self.filter_by_n_indexed(
-            self.all_solutions, n_indexed_cutoff=0.05
-        )  # 5 percent
-
-        if self.check_doubled_cell:
-            self.filtered_solutions = filter_doubled_cell(self.filtered_solutions)
-
-        self.filtered_solutions = self.filter_by_likelihood(self.filtered_solutions)
-
-        self.filtered_solutions = self.filter_by_volume(self.filtered_solutions)
-
-        self.filtered_solutions = self.filter_by_n_indexed(self.filtered_solutions)
-
-        return
-
-    def best_solution(self):
-        self.best_filtered_liklihood = max(
-            s.model_likelihood for s in self.filtered_solutions
-        )
-
-        solutions = [
-            s
-            for s in self.filtered_solutions
-            if s.model_likelihood == self.best_filtered_liklihood
-        ]
-        return solutions[0]
-
-    def __str__(self):
-        rows = []
-        rows.append(
-            ["unit_cell", "volume", "n_indexed", "fraction_indexed", "likelihood"]
-        )
-
-        for i, s in enumerate(self.all_solutions):
-            s = self.all_solutions[i]
-            rows.append(
-                [
-                    format(
-                        s.crystal.get_unit_cell(),
-                        "{:.2f} {:.2f} {:.2f} {:.1f} {:.1f} {:.1f}",
-                    ),
-                    "%.0f" % s.crystal.get_unit_cell().volume(),
-                    str(s.n_indexed),
-                    "%.0f" % (s.fraction_indexed * 100),
-                    "%.2f" % s.model_likelihood,
-                ]
-            )
-
-        from libtbx import table_utils
-
-        return table_utils.format(rows=rows, has_header=True)
-
-
-class SolutionTrackerWeighted(object):
-    def __init__(self, power=2, volume_weight=1, n_indexed_weight=1, rmsd_weight=1):
-        self.volume_weight = volume_weight
-        self.n_indexed_weight = n_indexed_weight
-        self.rmsd_weight = rmsd_weight
-        self.power = power
-        self.all_solutions = []
-
-    def append(self, item):
-        self.all_solutions.append(item)
-
-    def __len__(self):
-        return len(self.all_solutions)
-
-    def score_by_volume(self, reverse=False):
-        # smaller volume = better
-        volumes = flex.double(
-            s.crystal.get_unit_cell().volume() for s in self.all_solutions
-        )
-        score = flex.log(volumes) / math.log(2)
-        return self.volume_weight * (score - flex.min(score))
-
-    def score_by_rmsd_xy(self, reverse=False):
-        # smaller rmsds = better
-        rmsd_x, rmsd_y, rmsd_z = flex.vec3_double(
-            s.rmsds for s in self.all_solutions
-        ).parts()
-        rmsd_xy = flex.sqrt(flex.pow2(rmsd_x) + flex.pow2(rmsd_y))
-        score = flex.log(rmsd_xy) / math.log(2)
-        return self.rmsd_weight * (score - flex.min(score))
-
-    def score_by_rmsd_z(self, reverse=False):
-        # smaller rmsds = better
-        rmsd_x, rmsd_y, rmsd_z = flex.vec3_double(
-            s.rmsds for s in self.all_solutions
-        ).parts()
-        score = flex.log(rmsd_z) / math.log(2)
-        return score - flex.min(score)
-
-    def score_by_fraction_indexed(self, reverse=False):
-        # more indexed reflections = better
-        fraction_indexed = flex.double(s.fraction_indexed for s in self.all_solutions)
-        score = flex.log(fraction_indexed) / math.log(2)
-        return self.n_indexed_weight * (-score + flex.max(score))
-
-    def best_solution(self):
-        scores = self.solution_scores()
-        perm = flex.sort_permutation(scores)
-        return self.all_solutions[perm[0]]
-
-    def solution_scores(self):
-        scores = sum(
-            flex.pow(score.as_double(), self.power)
-            for score in (
-                self.score_by_fraction_indexed(),
-                self.score_by_volume(),
-                self.score_by_rmsd_xy(),
-                # self.score_by_rmsd_z(),
-            )
-        )
-        return scores
-
-    def __str__(self):
-        rows = []
-        rows.append(
-            [
-                "unit_cell",
-                "volume",
-                "volume score",
-                "#indexed",
-                "% indexed",
-                "% indexed score",
-                "rmsd_xy",
-                "rmsd_xy score",
-                #'rmsd_z', 'rank_rmsd_z',
-                "overall score",
-            ]
-        )
-
-        score_by_fraction_indexed = self.score_by_fraction_indexed()
-        score_by_volume = self.score_by_volume()
-        score_by_rmsd_xy = self.score_by_rmsd_xy()
-        score_by_rmsd_z = self.score_by_rmsd_z()
-        overall_scores = self.solution_scores()
-
-        perm = flex.sort_permutation(overall_scores)
-
-        rmsd_x, rmsd_y, rmsd_z = flex.vec3_double(
-            s.rmsds for s in self.all_solutions
-        ).parts()
-        rmsd_xy = flex.sqrt(flex.pow2(rmsd_x) + flex.pow2(rmsd_y))
-        rmsd_z *= 180 / math.pi
-
-        for i in perm:
-            s = self.all_solutions[i]
-            rows.append(
-                [
-                    format(
-                        s.crystal.get_unit_cell(),
-                        "{:.2f} {:.2f} {:.2f} {:.1f} {:.1f} {:.1f}",
-                    ),
-                    "%.0f" % s.crystal.get_unit_cell().volume(),
-                    "%.2f" % score_by_volume[i],
-                    str(s.n_indexed),
-                    "%.0f" % (s.fraction_indexed * 100),
-                    "%.2f" % score_by_fraction_indexed[i],
-                    "%.2f" % rmsd_xy[i],
-                    "%.2f" % score_by_rmsd_xy[i],
-                    # "%.2f" %rmsd_z[i], "%.2f" %score_by_rmsd_z[i],
-                    "%.2f" % overall_scores[i],
-                ]
-            )
-
-        from libtbx import table_utils
-
-        return table_utils.format(rows=rows, has_header=True)
 
 
 def optimise_basis_vectors(reciprocal_lattice_points, vectors):
